TaskId: TRE-234-FE (16) Implement delivery address route feature
Type: Development

Description:
- Update api endpoint
