+++ app/order-ui/src/api/index.ts
@@ -23,4 +23,3 @@ export * from './voucher'
 export * from './gift-card'
 export * from './card-order'
 export * from './balance'
-export * from './google-map'
