+++ app/order-ui/src/hooks/index.ts
@@ -30,4 +30,3 @@ export * from './use-card-order'
 export * from './use-balance'
 export * from './use-point-transaction'
 export * from './use-payment'
-export * from './use-google-map'
