+++ app/order-ui/src/hooks/use-order.ts
@@ -26,6 +26,10 @@ import {
   getOrderProvisionalBill,
   reprintFailedInvoicePrinterJobs,
   updatePublicVoucherInOrder,
+  getAddressDirection,
+  getDistanceAndDuration,
+  getAddressByPlaceId,
+  getAddressSuggestions,
 } from '@/api'
 import {
   ICreateOrderRequest,
@@ -39,6 +43,7 @@ import {
   IUpdateNoteRequest,
   IOrderItemsParam,
 } from '@/types'
+import { QUERYKEY } from '@/constants'
 
 export const useOrders = (q: IOrdersQuery) => {
   return useQuery({
@@ -280,5 +285,45 @@ export const useReprintFailedInvoicePrinterJobs = () => {
     mutationFn: async (slug: string) => {
       return reprintFailedInvoicePrinterJobs(slug)
     },
+  })
+}
+
+export const useGetAddressSuggestions = (address: string) => {
+  return useQuery({
+    queryKey: [QUERYKEY.addressSuggestions, address],
+    queryFn: () => getAddressSuggestions(address),
+    enabled: !!address,
+  })
+}
+
+export const useGetAddressByPlaceId = (placeId: string) => {
+  return useQuery({
+    queryKey: [QUERYKEY.addressByPlaceId, placeId],
+    queryFn: () => getAddressByPlaceId(placeId),
+    enabled: !!placeId,
+  })
+}
+
+export const useGetAddressDirection = (
+  branch: string,
+  lat: number,
+  lng: number,
+) => {
+  return useQuery({
+    queryKey: [QUERYKEY.addressDirection, branch, lat, lng],
+    queryFn: () => getAddressDirection(branch, lat, lng),
+    enabled: !!branch && !!lat && !!lng,
+  })
+}
+
+export const useGetDistanceAndDuration = (
+  branch: string,
+  lat: number,
+  lng: number,
+) => {
+  return useQuery({
+    queryKey: [QUERYKEY.distanceAndDuration, branch, lat, lng],
+    queryFn: () => getDistanceAndDuration(branch, lat, lng),
+    enabled: !!branch && !!lat && !!lng,
   })
 }
